+++ src/elements/Alert/Warning.jsx
@@ -19,7 +19,6 @@ const Warning = React.createClass({
                 {children}
             </Alert>
         );
-
     },
 });