+++ src/mixins/PerformanceMixin.js
@@ -1,27 +1,35 @@
 import _ from 'lodash';
 
 const getDisplayName = function(Component) {
-    if (__DEBUG__) {
-        const objectConstructor = Object.getPrototypeOf(Component).constructor;
-        const objectTest = (typeof objectConstructor !== 'undefined') ? objectConstructor : Component;
-        return objectTest.displayName || objectTest.name ||
-               (typeof objectTest === 'string' ? objectTest : objectTest.toString());
-    }
+    const objectConstructor = Object.getPrototypeOf(Component).constructor;
+    const objectTest =
+        typeof objectConstructor !== 'undefined'
+            ? objectConstructor
+            : Component;
+    return (
+        objectTest.displayName ||
+        objectTest.name ||
+        (typeof objectTest === 'string' ? objectTest : objectTest.toString())
+    );
 };
 
 const showDifferences = function(object, config, nextConfig) {
     if (__DEBUG__) {
         const componentName = getDisplayName(object);
-        const diff = _.reduce(object[config], function(result, value, key) {
-            return _.isEqual(value, nextConfig[key]) ? result : result.concat(key);
-        }, []);
-        console.log(componentName + '.' + config + ' not equal for: ' + diff.join(', '));
+        const diff = _.reduce(
+            object[config],
+            (result, value, key) =>
+                _.isEqual(value, nextConfig[key]) ? result : result.concat(key),
+            []
+        );
+        console.log(
+            `${componentName}.${config} not equal for: ${diff.join(', ')}`
+        );
     }
-}
+};
 
 const PerformanceMixin = {
     shouldComponentUpdate(nextProps, nextState) {
-
         if (_.isEqual(nextState, this.state) === false) {
             if (__DEBUG__ && window.enablePerformanceMixingLog) {
                 showDifferences(this, 'state', nextState);
@@ -38,7 +46,6 @@ const PerformanceMixin = {
 
         return false;
     },
-
 };
 
 export default PerformanceMixin;